linkerscipt_ARM: Fix entry addresses for ARM architectures
The entry address for ARM architectures is calculated at
arm_setup_direct_kernel_boot() in hw/arm/boot.c

At a high level, if it finds an ELF it assumes is baremetal,
if instead it finds a RAW binary it assumes its a Linux kernel.

For ELF files it parses the ELF header and gets the entry address
from there.

For RAW binaries, it adds an offset based on RAM:
*entry = mem_base + kernel_load_offset

This offset is:
KERNEL_LOAD_ADDR = 0x1000 for 32 bit ARM
and
KERNEL64_LOAD_ADDR = 0x80000 for 64 bit ARM.

In load_aarch64_image, regardless of the entry address specified
in the linkerscript, it relocates the binary to the entry address
defined by mem_base + kernel_load_offset (this function does not
get executed for ELF files).

This means breakpoints wont work since the symbol table has a
different address than the one the binary is actually loaded at,
but the program will run.

This approach from QEMU is smarter than expected, a bit fool-proof,
which makes things easier for us but it can also easily hide bugs
in this code, just like the one I intend to fix here.

Set the proper ENTRY addresses for ARM architectures given by the
offsets mentioned above.

Additionally fix comment about stack size for aarch64.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandro@enedino.org>