Makefile: Fix race condition when creating build directory
There have been a couple of instances where the CI is reporting
an error whilst assembling the code, these have been happening
randomly in a series of architectures.

It appears that the build directory is not being created on time,
when the assembler is trying to create its output files the
directory is still not there.

Modify the builddir dependencies to use order-only prerequisite,
as specified in the GNU Make manual [1].

[1] https://www.gnu.org/software/make/manual/make.html#Prerequisite-Types

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandro@enedino.org>